refactor: move top functions to own file
